refactor: split large handlers into reusable modules
- extract file discovery/selection into handlers/file_selection_mixin.py\n- extract shared text beautification helpers into utils/shared/text_formatting.py\n- split mobile static scan internals into constants/helpers modules\n- keep compatibility through thin static scan mixin wrappers\n- update structure.txt with per-file summaries and aligned inline comments
