+++ .gitignore
@@ -22,3 +22,4 @@ dependencies.txt
 *.log
 *.pyc
 *.pyo
+.venv
