+++ utils/internal/password_constants.py
@@ -1,0 +1,3 @@
+"""Constants for password module workflows."""
+
+PASSWORD_OUTPUT_EXTENSION = "txt"
