+++ utils/mobile/__init__.py
@@ -1,4 +1,4 @@
-
 from .mobile_commands import MobileCommands
+from .models import Finding
 
-__all__ = [ MobileCommands]
+__all__ = ["MobileCommands", "Finding"]
